pyproject.toml
[tool.poetry]
name = "tinfit1"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
flask = "^1.1.2"
Flask-WTF = "^0.14.3"
Flask-Bootstrap4 = "^4.0.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
